Flushes BufWriter
It seems important to do because the Rust Documentation says :
"It is critical to call flush before BufWriter<W> is dropped. Though dropping will attempt to flush the contents of the buffer, any errors that happen in the process of dropping will be ignored. Calling flush ensures that the buffer is empty and thus dropping will not even attempt file operations."
Rust documentation link : https://doc.rust-lang.org/std/io/struct.BufWriter.html